removed 'base' line search, bracketing line search is now the base. avoids confusion about having incomplete 'base' modules which aren't useful on their own
