+++ seisflows/plugins/line_search/bracket.py
@@ -1,25 +1,28 @@
 #!/usr/bin/env python3
 """
-This is the subclass class for seisflows.plugins.line_search.bracket
+This is the Bracketing line search class for seisflows
+
+Line search is called on by the optimization procedure and should not really
+have any agency (i.e. it should not be able to iterate its own step count etc.,
+this should be completely left to the optimization algorithm to keep everything
+in one place)
 """
 import logging
 import numpy as np
 
-from seisflows.tools import msg
-from seisflows.plugins.line_search.base import Base
+from seisflows.tools.array import count_zeros
 from seisflows.tools.math import parabolic_backtrack, polynomial_fit
 
 
-class Bracket(Base):
+class Bracket:
     """
-    Implements bracketing line search, which attempts to find a step length
-    corresponding to misfit reduction, and a misfit corresponding to misfit
-    increase, so that the optimization procedure can scale the step length
-    in future iterations.
+    Abstract base class for line search
 
     Variables Descriptions:
         x: list of step lenths from current line search
         f: correpsonding list of function values
+        m: number of step lengths in current line search
+        n: number of model updates in optimization problem
         gtg: dot product of gradient with itself
         gtp: dot product of gradient and search direction
 
@@ -31,12 +34,149 @@ class Bracket(Base):
     # Class-specific logger accessed using self.logger
     logger = logging.getLogger(__name__).getChild(__qualname__)
 
-    def __init__(self, **kwargs):
+    def __init__(self):
         """
-        These parameters should not be set by the user.
-        Attributes are initialized as NoneTypes for clarity and docstrings.
+
+        :type step_count_max: int
+        :param step_count_max: maximum number of step counts before changing
+            line search behavior. set by PAR.STEP_COUNT_MAX
+        :type step_len_max: int
+        :param step_len_max: maximum length of the step, defaults to infinity,
+            that is unbounded step length. set by PAR.STEP_LEN_MAX
+        :type log_file: str
+        :param log_file: path to write line search stats to. set by optimize.setup()
         """
-        super().__init__(**kwargs)
+        # Prepare lists for line search history
+        self.func_vals = []
+        self.step_lens = []
+        self.gtg = []
+        self.gtp = []
+        self.step_count = 0
+
+    def initialize(self, step_len, func_val, gtg, gtp):
+        """
+        Initialize a new search from step count 0 and calculate the step
+        direction and length
+
+        :type iter: int
+        :param iter: current iteration defined by OPTIMIZE.iter
+        :type step_len: float
+        :param step_len: initial step length determined by optimization
+        :type func_val: float
+        :param func_val: current evaluation of the objective function
+        :type gtg: float
+        :param gtg: dot product of the gradient with itself
+        :type gtp: float
+        :param gtp: dot product of gradient `g` with search direction `p`
+        :rtype alpha: float
+        :return alpha: the calculated trial step length
+        :rtype status: int
+        :return status: current status of the line search
+        """
+        self.step_count = 0
+        self.step_lens += [step_len]
+        self.func_vals += [func_val]
+        self.gtg += [gtg]
+        self.gtp += [gtp]
+
+        # Call calculate step, must be implemented by subclass
+        alpha, status = self.calculate_step()
+
+        return alpha, status
+
+    def update(self, step_len, func_val):
+        """
+        Update search history by appending internal attributes, writing the
+        current list of step lengths and function evaluations, and calculating a
+        new step length
+
+        :type iter: int
+        :param iter: current iteration defined by OPTIMIZE.iter
+        :type step_len: float
+        :param step_len: step length determined by optimization
+        :type func_val: float
+        :param func_val: current evaluation of the objective function
+        :rtype alpha: float
+        :return alpha: the calculated rial step length
+        :rtype status: int
+        :return status: current status of the line search
+        """
+        # This has been moved into workflow.line_search()
+        # self.step_count += 1
+        self.step_lens += [step_len]
+        self.func_vals += [func_val]
+
+        # Call calcuate step, must be implemented by subclass
+        alpha, status = self.calculate_step()
+
+        return alpha, status
+
+    def clear_history(self):
+        """
+        Clears internal line search history
+        """
+        self.func_vals = []
+        self.step_lens = []
+        self.gtg = []
+        self.gtp = []
+        self.step_count = 0
+
+    def reset(self):
+        """
+        If a line search fails mid-search, and the User wants to resume from 
+        the line search function. Initialize will be called again. This function
+        undos the progress made by the previous line search so that a new line
+        search can be called without problem.
+
+        output.optim needs to have its lines cleared manually
+        """
+        # First step treated differently
+        if len(self.step_lens) <= 1:
+            self.clear_history()
+        else:
+            # Wind back dot products by one
+            self.gtg = self.gtg[:-1]
+            self.gtp = self.gtp[:-1]
+            
+            # Move step lens and function evaluations by number of step count
+            original_idx = -1 * self.step_count - 1
+            self.step_lens = self.step_lens[:original_idx]
+            self.func_vals = self.func_vals[:original_idx]
+
+    def search_history(self, sort=True):
+        """
+        A convenience function, collects information based on the current
+        evaluation of the line search, needed to determine search status and 
+        calculate step length. From the full collection of the search history,
+        only returns values relevant to the current line search.
+
+        :type sort: bool
+        :param sort: sort the search history by step length
+        :rtype x: np.array
+        :return x: list of step lenths from current line search
+        :rtype f: np.array
+        :return f: correpsonding list of function values
+        :rtype gtg: list
+        :return gtg: dot product dot product of gradient with itself
+        :rtype gtp: list
+        :return gtp: dot product of gradient and search direction
+        :rtype i: int
+        :return i: step_count
+        :rtype j: int
+        :return j: number of iterations corresponding to 0 step length
+        """
+        i = self.step_count
+        j = count_zeros(self.step_lens) - 1
+        k = len(self.step_lens)
+        x = np.array(self.step_lens[k - i - 1:k])
+        f = np.array(self.func_vals[k - i - 1:k])
+
+        # Sort by step length
+        if sort:
+            f = f[abs(x).argsort()]
+            x = x[abs(x).argsort()]
+
+        return x, f, self.gtg, self.gtp, i, j
 
     def calculate_step(self):
         """
@@ -51,7 +191,7 @@ class Bracket(Base):
         f_str = ", ".join([f"{_:.2E}" for _ in f])
         self.logger.debug(f"step length(s) = {x_str}")
         self.logger.debug(f"misfit val(s)  = {f_str}")
-        
+
         # For the first inversion and initial step, set alpha manually
         if step_count == 0 and update_count == 0:
             # Based on idea from Dennis and Schnabel
@@ -66,12 +206,12 @@ class Bracket(Base):
             self.logger.info(f"first step, setting scaled step length")
             status = 0
         # If misfit is reduced and then increased, we've bracketed. Pass
-        elif self._check_bracket(x, f) and self._good_enough(x,f):
+        elif _check_bracket(x, f) and _good_enough(x, f):
             alpha = x[f.argmin()]
             self.logger.info(f"bracket okay, step length reasonable, pass")
             status = 1
         # If misfit is reduced but not close, set to quadratic fit
-        elif self._check_bracket(x, f):
+        elif _check_bracket(x, f):
             alpha = polynomial_fit(x, f)
             self.logger.info(f"bracket okay, step length unreasonable, "
                              f"manual step")
@@ -111,52 +251,52 @@ class Bracket(Base):
 
         return alpha, status
 
-    @staticmethod
-    def _check_bracket(step_lens, func_vals):
-        """
-        Checks if minimum has been bracketed
 
-        Looks at the minimum of the misfit values calculated through eval func
-        to see if the misfit has been reduced w.r.t the initial misfit
+def _check_bracket(step_lens, func_vals):
+    """
+    Checks if minimum has been bracketed
 
-        :type step_lens: numpy.array
-        :param step_lens: an array of the step lengths taken during iteration
-        :type func_vals: numpy.array
-        :param func_vals: array of misfit values from eval func function
-        :rtype: bool
-        :return: status of function as a bool
-        """
-        x, f = step_lens, func_vals
-        imin, fmin = f.argmin(), f.min()
-        if (fmin < f[0]) and any(f[imin:] > fmin):
+    Looks at the minimum of the misfit values calculated through eval func
+    to see if the misfit has been reduced w.r.t the initial misfit
+
+    :type step_lens: numpy.array
+    :param step_lens: an array of the step lengths taken during iteration
+    :type func_vals: numpy.array
+    :param func_vals: array of misfit values from eval func function
+    :rtype: bool
+    :return: status of function as a bool
+    """
+    x, f = step_lens, func_vals
+    imin, fmin = f.argmin(), f.min()
+    if (fmin < f[0]) and any(f[imin:] > fmin):
+        okay = True
+    else:
+        okay = False
+    return okay
+
+def _good_enough(step_lens, func_vals, thresh=np.log10(1.2)):
+    """
+    Checks if step length is reasonably close to quadratic estimate
+
+    :type step_lens: np.array
+    :param step_lens: an array of the step lengths taken during iteration
+    :type func_vals: np.array
+    :param func_vals: array of misfit values from eval func function
+    :type thresh: numpy.float64
+    :param thresh: threshold value for comparison against quadratic estimate
+    :rtype: bool
+    :return: status of function as a bool
+    """
+    x, f = step_lens, func_vals
+    if not _check_bracket(x, f):
+        okay = False
+    else:
+        x0 = polynomial_fit(x, f)
+        if any(np.abs(np.log10(x[1:] / x0)) < thresh):
             okay = True
         else:
             okay = False
-        return okay
-
-    def _good_enough(self, step_lens, func_vals, thresh=np.log10(1.2)):
-        """
-        Checks if step length is reasonably close to quadratic estimate
-
-        :type step_lens: np.array
-        :param step_lens: an array of the step lengths taken during iteration
-        :type func_vals: np.array
-        :param func_vals: array of misfit values from eval func function
-        :type thresh: numpy.float64
-        :param thresh: threshold value for comparison against quadratic estimate
-        :rtype: bool
-        :return: status of function as a bool
-        """
-        x, f = step_lens, func_vals
-        if not self._check_bracket(x, f):
-            okay = False
-        else:
-            x0 = polynomial_fit(x, f)
-            if any(np.abs(np.log10(x[1:] / x0)) < thresh):
-                okay = True
-            else:
-                okay = False
-        return okay
+    return okay
